first pass at renaming
