+++ Makefile
@@ -48,7 +48,7 @@ all: default
 SOURCES += src/algorithms.c src/amy.c src/envelope.c src/examples.c src/parse.c \
 	src/filters.c src/oscillators.c src/pcm.c src/partials.c src/interp_partials.c src/custom.c \
 	src/delay.c src/log2_exp2.c src/patches.c src/transfer.c src/sequencer.c \
-	src/libminiaudio-audio.c src/instrument.c src/midi.c
+	src/libminiaudio-audio.c src/instrument.c src/midi.c src/api.c
 
 OBJECTS = $(patsubst %.c, %.o, $(SOURCES))
