+++ src/pyamy.c
@@ -13,7 +13,7 @@ static PyObject * send_wrapper(PyObject *self, PyObject *args) {
     if (! PyArg_ParseTuple(args, "s", &arg1)) {
         return NULL;
     }
-    amy_play_message(arg1);
+    amy_add_message(arg1);
     return Py_None;
 }
